refactor: schema helpers
