+++ src/components/RenderForm.js
@@ -5,7 +5,7 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 import { customizeValidator } from '@rjsf/validator-ajv8'
 import { widgets } from '../custom-ui/CustomWidgets'
 import { uiSchema } from '../custom-ui/CustomUISchema'
-import { AJV_OPTIONS } from '../utils/helpers/schema-handlers.helpers'
+import Config from '../utils/config'
 import { saveToJSONFile } from '../utils/helpers/file.helpers'
 import { toast } from 'react-toastify'
 
@@ -21,7 +21,7 @@ function RenderForm (props) {
   */
   const { schemaType, schema, formData } = props
   const formRef = createRef()
-  const validator = customizeValidator(AJV_OPTIONS)
+  const validator = customizeValidator(Config.ajvOptions)
 
   /**
    * Event handler to validate the form
